Revert "Status update split out"
This reverts commit fa8fb398a962137273d30cdc392b69ae7f54ed62.
