--- UserInterface.py
@@ -1,25 +1,0 @@
-from _thread import *
-
-class UserInterface:
-
-    def __init__(self):
-        pass
-
-    def setStartRace(func):
-        pass
-
-    def setStopRace(func):
-        pass
-    
-    def getRaceState(self):
-        return 0
-
-    def getRaceData(self):
-        pass
-    
-    def main(self):
-        pass
-
-
-    
-
